Reorganized Code
Moved "learning" stuff from Agent.py to main.py, so agent.py is no longer needed. Cleaned up and reorganized "learning" stuff to be less complicated and more sensible.

Reorganized game2048.py so that keyboard bindings are only generated when the file itself is executed. Changed timings of when screen is painted.